feat: clear meta in chunk
